django-backend/requirements.txt
arrow==0.15.8
asgiref==3.2.10
binaryornot==0.4.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
Django==3.1
django-cors-headers==3.4.0
django-extensions==3.0.5
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-filters==1.0.0.dev0
drf-flex-fields==0.8.5
drf-nested-routers==0.91
idna==2.10
Jinja2==2.11.2
jinja2-time==0.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
poyo==0.5.0
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.10
